move template and options to its own files
